Drop mod_ file prefix for tests exercising code in mod.ts files. Merge unit tests that were split out between unwrapped and wrapped typed objects.
